requirements.txt
hyperliquid-python-sdk>=0.22.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
websocket-client>=1.6.0
eth-account>=0.9.0
matplotlib>=3.7.0
